+++ packages/banana-react/src/index.ts
@@ -10,9 +10,12 @@ import { Marquee } from './marquee';
 import { Menu } from './menu';
 import { MenuItem } from './menu-item';
 import { Message } from './message';
+import { Modal } from './modal';
 import { Overlay } from './overlay';
 import { Popup } from './popup';
 import { Progress } from './progress';
+import { Radio } from './radio';
+import { RadioGroup } from './radio-group';
 import { Rating } from './rating';
 import { Select } from './select';
 import { SelectOption } from './select-option';
@@ -32,9 +35,12 @@ export {
   Menu,
   MenuItem,
   Message,
+  Modal,
   Overlay,
   Popup,
   Progress,
+  Radio,
+  RadioGroup,
   Rating,
   Select,
   SelectOption,